requirements.txt
fastapi
python-telegram-bot[aio]==22.4
uvicorn
requests
gunicorn
